+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Changed AdaptivePolicyGroup CreatedAt and UpdatedAt to be DateTime
 - Changed AdaptivePolicyPolicies to correct nullability. Changed timestamps to DateTime
 - Updated RoutingOspf to add RoutingOspfV3 settings
+- Made all DeviceClient properties null
 
 ## v1.20.42
 - v1.15 changes